requirements-mac.txt
pandas
typing
transformers
pycryptodome
secretstorage
nltk
openai